+++ 05-JWT-Basics/starter/middleware/auth.js
@@ -1,5 +1,20 @@
+const jwt = require('jsonwebtoken')
+const CustomAPIError = require('../errors/custom-error')
+require('dotenv').config
+
 const authorizationMiddleware = async (req, res, next) => {
-  console.log(req.headers.authorization)
-  next()
+  const authHeaders = req.headers.authorization
+  if (!authHeaders || !authHeaders.startsWith('Bearer ')) {
+    throw new CustomAPIError('No token provided', 401)
+  }
+  const token = authHeaders.split(' ')[1]
+  try {
+    const decoded = jwt.verify(token, process.env.JWT_SECRET)
+    const { id, username } = decoded
+    req.user = { id, username }
+    next()
+  } catch (error) {
+    throw new CustomAPIError('Not authorize to access this route', 401)
+  }
 }
 module.exports = authorizationMiddleware
